android: minor refactor disk shadercache
